repositories Error handling refactoring

+++ internal/service/post/interface.go
@@ -7,8 +7,8 @@ import (
 
 type Service interface {
 	CreatePost(ctx context.Context, post *model.CreatePostDTO) (*model.Post, error)
-	GetPostByID(ctx context.Context, id int) (*model.PostDetailed, error)
+	GetPostByID(ctx context.Context, id int64) (*model.PostDetailed, error)
 	ListPosts(ctx context.Context, filters *model.PostFilters) ([]*model.PostDetailed, error)
-	UpdatePost(ctx context.Context, id int, post *model.UpdatePostDTO) error
-	DeletePost(ctx context.Context, id int) error
+	UpdatePost(ctx context.Context, id int64, post *model.UpdatePostDTO) error
+	DeletePost(ctx context.Context, id int64) error
 }
